Fix extension
