+++ Korenko HM 24/Korenko HM 24/CellForMainTableView/HourlyWeatherCell/CellForCollectionView/CollectionCellForHourly.swift
@@ -27,8 +27,8 @@ class CollectionCellForHourly: UICollectionViewCell {
         DispatchQueue.main.async { [weak self] in
             guard let self = self,
                   let weather = weatherData.weather.first else { return }
-            self.timeLabel.text = self.confertUnixTypeToNormal(time: Double(weatherData.dt), typeTime: "HH:mm")
-            self.iconImageView.image = self.getIconImage(iconId: weather.icon)
+            self.timeLabel.text = Double(weatherData.dt).timeHHmm
+            self.iconImageView.image = weather.icon.image
             self.tempLabel.text = "\(Int(weatherData.temp)) C"
         }
     }
